+++ src/components/BackToTop/tests/BackToTop.test.js
@@ -16,7 +16,7 @@ describe('BackToTop', () => {
 
     useLayoutEffect(() => {
       test('it forwards the given ref correctly', () => {
-        expect(ref.current).not.toBeFalsy()
+        expect(ref.current instanceof HTMLButtonElement).toBeTruthy()
       })
     }, [ref])