+++ src/components/CookiesBanner/tests/CookiesBanner.test.js
@@ -22,7 +22,7 @@ describe('CookiesBanner', () => {
 
     useLayoutEffect(() => {
       test('it forwards the given ref correctly', () => {
-        expect(ref.current).not.toBeFalsy()
+        expect(ref.current instanceof HTMLDivElement).toBeTruthy()
       })
     }, [ref])