+++ src/components/CreditCardNumber/tests/CreditCardNumber.test.js
@@ -10,7 +10,7 @@ describe('CrediCardNumber', () => {
 
     useLayoutEffect(() => {
       test('it forwards the given ref correctly', () => {
-        expect(ref.current).not.toBeFalsy()
+        expect(ref.current instanceof HTMLInputElement).toBeTruthy()
       })
     }, [ref])