+++ src/components/Navbar/tests/Navbar.test.js
@@ -14,7 +14,7 @@ describe('Navbar', () => {
 
     useLayoutEffect(() => {
       test('it forwards the given ref correctly', () => {
-        expect(ref.current).not.toBeFalsy()
+        expect(ref.current instanceof HTMLElement).toBeTruthy()
       })
     }, [ref])